add confirm dialog for update/delete

+++ src/app/user/user-posts/user-post-card/user-post-card.component.ts
@@ -1,6 +1,8 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { PostService } from 'src/app/post/posts.service';
+import { AppYesNoDialogComponent } from 'src/app/shared/dialogs/app-yes-no-dialog/app-yes-no-dialog.component';
 import { PostStatusEnum } from 'src/app/shared/enums/PostStatusEnum';
 import { Post } from 'src/app/shared/models/post.model';
 
@@ -14,7 +16,9 @@ export class UserPostCardComponent implements OnInit {
   @Output() loading = new EventEmitter<Number>();
 
   imageShown = true;
-  constructor(private router: Router, private postService: PostService) { }
+  constructor(private router: Router,
+              private postService: PostService,
+              private dialog: MatDialog) { }
 
   ngOnInit(): void {
     console.log(this.post);
@@ -28,31 +32,55 @@ export class UserPostCardComponent implements OnInit {
     this.imageShown = false;
   }
 
-  /**
-   * 1. Delete the post from the database
-   * 2. Refetch the user list of posts
-   * 3. Notify the user post page that a post is deleted
-   */
-  onDeletePost(){
-    this.loading.emit(this.post.status);
+  onOpenConfirmDialog(msg){
+    const dialogRef = this.dialog.open(AppYesNoDialogComponent, {
+      data : {
+        message: msg
+      }
+    });
 
-    this.postService.deleteUserPostFromDb(this.post.id);
-
-    this.postService.fetchUserPosts();
+    return dialogRef;
   }
 
   /**
-   * 1. Update Post with status=Done
-   * 2. Refetch the user list of posts
-   * 3. Notify the user post page that a post is updated
+   * 1. Open the confirm dialog
+   * 2. Notify the user post page that a post is deleted
+   * 3. Delete the post from the database
+   * 4. Refetch the user list of posts
+   * 5. Refetch the list of active posts
+   */
+  onDeletePost(){
+    this.onOpenConfirmDialog("This action will cause you no longer have access to it. Are you sure you want to delete it?")
+      .afterClosed().subscribe((result)=>
+        {
+          if (result==true){
+
+            this.loading.emit(this.post.status);
+
+            this.postService.deleteUserPostFromDb(this.post.id);
+          }
+        }
+      )
+  }
+
+  /**
+   * 1. Open the confirm dialog
+   * 2. Notify the user post page that a post is updated
+   * 3. Update Post with status=Done
+   * 4. Refetch the user list of posts
+   * 3. refetch the active posts
    */
   onMarkPostDone(){
-    this.loading.emit(this.post.status);
+    this.onOpenConfirmDialog("This action will remove the post from home page. Other users will no longer have access to it. Are you sure that you want to perform this action?")
+      .afterClosed()
+      .subscribe((result)=>{
+        if (result==true){
+          this.loading.emit(this.post.status);
 
-    this.post.status = PostStatusEnum.Done;
+          this.post.status = PostStatusEnum.Done;
 
-    this.postService.updateUserPostInDb(this.post);
-
-    this.postService.fetchUserPosts();
+          this.postService.updateUserPostInDb(this.post);
+        }
+      })
   }
 }
